updates to pytorch tests, encompassing pr 71

+++ tests/unit/pytorch/models/test_continuous_model.py
@@ -150,7 +150,7 @@ def test_ContinuousModel(plot):
 
     # predictive distribution covered for each sample
     cov = model.pred_dist_coverage(x[:11, :], y[:11, :])
-    assert isinstance(cov, np.float)
+    assert isinstance(cov, float)
 
     # plot coverage by
     xo, co = model.coverage_by(x[:, :1], x, y)
